fix: move schema out of commands folder
