+++ src/cli/schema/device.ts
